+++ plugins/css-atrule.js
@@ -1,6 +1,7 @@
 /*
     处理样式规则中的@规则
  */
+let util = require('util');
 let regexp = require('./util-rcache');
 let {
     genCssSelector
@@ -16,17 +17,20 @@ let keyframesReg = /(^|[\s\}])(@(?:-webkit-|-moz-|-o-|-ms-)?keyframes)\s+(['"])?
 let genCssContentReg = key => {
     return regexp.get('\\b(font-family|animation|animation-name)\\s*:([^\\{\\}:\\r\\n\\(\\)]*?)([\'"])?' + regexp.escape(key) + '\\3(?=[,\\s;])', 'g');
 };
+let globalContents = {};
 //css @规则的处理
-module.exports = (fileContent, cssNamesKey) => {
+let processor = (fileContent, cssNamesKey, addToGlobal) => {
     let contents = [];
     //先处理keyframes
     fileContent = fileContent.replace(keyframesReg, (m, head, keyframe, q, name) => {
         //把名称保存下来，因为还要修改使用的地方
-        contents.push(name);
-        name = genCssSelector(name, cssNamesKey);
+        if (contents.indexOf(name) == -1) {
+            contents.push(name);
+        }
+        let tname = genCssSelector(name, cssNamesKey);
         q = q || '';
         //增加前缀
-        return head + keyframe + ' ' + q + name + q;
+        return head + keyframe + ' ' + q + tname + q;
     });
     //处理其它@规则，这里只处理了font-face
     fileContent.replace(fontfaceReg, (match, content) => {
@@ -45,18 +49,43 @@ module.exports = (fileContent, cssNamesKey) => {
             if (parts.length && parts[0].trim() === 'font-family') {
                 let fname = parts[1].trim();
                 fname = fname.replace(trimQ, '');
-                contents.push(fname);
+                if (contents.indexOf(fname) == -1) {
+                    contents.push(fname);
+                }
                 break;
             }
         }
     });
+    for (let p in globalContents) {
+        if (globalContents.hasOwnProperty(p)) {
+            if (contents.indexOf(p) == -1) {
+                contents.push({
+                    t: p,
+                    tn: globalContents[p]
+                });
+            }
+        }
+    }
     //contents中目前只有@font-face及@keyframes2种
     while (contents.length) {
-        let t = contents.pop();
-        let reg = genCssContentReg(t);
-        t = genCssSelector(t, cssNamesKey);
-        fileContent = fileContent.replace(reg, '$1:$2$3' + t + '$3');
+        let t = contents.pop(),
+            reg, tn;
+        if (util.isString(t)) {
+            reg = genCssContentReg(t);
+            tn = genCssSelector(t, cssNamesKey);
+            if (addToGlobal) {
+                globalContents[t] = tn;
+            }
+        } else {
+            reg = genCssContentReg(t.t);
+            tn = t.tn;
+        }
+        fileContent = fileContent.replace(reg, '$1:$2$3' + tn + '$3');
     }
     //console.log(fileContent);
     return fileContent;
 };
+processor.reset = () => {
+    globalContents = {};
+};
+module.exports = processor;