+++ plugins/util-init.js
@@ -9,8 +9,7 @@ module.exports = () => {
         configs.$inited = 1;
         configs.tmplFolder = path.resolve(configs.tmplFolder);
         configs.srcFolder = path.resolve(configs.srcFolder);
-        configs.compileFileExtNamesReg = new RegExp('\\.(?:' + configs.compileFileExtNames.join('|') + ')$');
-        configs.tmplFileExtNamesReg = new RegExp('\\.(?:' + configs.tmplFileExtNames.join('|') + ')$');
+        configs.jsFileExtNamesReg = new RegExp('\\.(?:' + configs.jsFileExtNames.join('|') + ')$');
         configs.moduleIdRemovedPath = configs.tmplFolder; //把路径中开始到模板目录移除就基本上是模块路径了
         if (!configs.disableMagixUpdater && !configs.tmplCommand) {
             configs.tmplCommand = /<%[\s\S]*?%>/g;
@@ -19,10 +18,7 @@ module.exports = () => {
             configs.tmplCommand = /<%[\s\S]*?%>/g;
         }
         if (!configs.cssSelectorPrefix) {
-            configs.cssSelectorPrefix = 'mx' + md5(configs.tmplFolder, 3, 'md5CssFileLen');
-        }
-        if (configs.debug) {
-            configs.compressCss = false;
+            configs.cssSelectorPrefix = 'x' + md5(configs.tmplFolder, 3, 'md5CssFileLen');
         }
     }
 };